aarch64-prep: Define a post kbuild hook for qemu-aarch64-system-virt
This hook packages the kernel image for u-boot using mkimage. It uses
KLOADADDR and KENTRYADDR for this.

The hook also uses mkimage to create a u-boot bootscript image. The
script is set up to populate u-boot's 'bootarg' environment variable
with the UUID of the filesystem image. In addition, the bootscript is
also set up to fetch the kernel image over ethernet and load it at
KLOADADDR followed by a jump to KENTRYADDR.

About bootarg: As it happens, u-boot appends the content of bootargs
into the device-tree image's 'chosen' node. Since we use DT for aarch64
kernels, this is a handy way to pass the UUID to the kernel. The
alternative would have been to directly populate the kernel's boot time
environment region with the UUID but that is bothersome to do in
assembly and hence this route.